rename: Tiltle component in garden page to Title
